+++ internal/tests/features/registry_register_lazy_service_test.go
@@ -22,14 +22,12 @@ func Test_Registry_register_lazy_service_type_factory(t *testing.T) {
 	ctx := resolving.NewScopedContext(context.Background())
 
 	// Act
-	lazyFactory, err := resolving.ResolveRequiredService[func() features.Lazy[*foo]](resolver, ctx)
-	actual := lazyFactory()
+	actual, err := resolving.ResolveRequiredService[features.Lazy[*foo]](resolver, ctx)
 	fooInstance0 := actual.Value()
 	fooInstance1 := actual.Value()
 
 	// Assert
 	assert.NoError(t, err)
-	assert.NotNil(t, lazyFactory)
 	assert.NotNil(t, actual)
 	assert.NotNil(t, fooInstance0)
 	assert.NotNil(t, fooInstance1)